Forced switch event added
